Revert "Created basic classes"
