+++ src/test/java/seedu/duke/DukeTest.java
@@ -1,17 +1,12 @@
-package seedu.duke;  //same package as the class being tested
+package seedu.duke;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-public class DukeTest {
+class DukeTest {
     @Test
-    public void dummyTest(){
-        assertEquals(2, 2);
-    }
-
-    @Test
-    public void anotherDummyTest(){
-        assertEquals(4, 4);
+    public void sampleTest() {
+        assertTrue(true);
     }
 }
